전체적인 로직 완성

+++ p2uhomepage/src/main/java/com/pray2you/p2uhomepage/domain/item/dto/request/CreateItemRequestDTO.java
@@ -17,7 +17,7 @@ public class CreateItemRequestDTO {
     @NotBlank
     private String itemName;
     @NotNull
-    private long point;
+    private int point;
     @NotBlank
     private String itemDescription;
